added UserInput method for interactiion with the Menu options. the tutorial i followed is: https://www.youtube.com/watch?v=LcCpqDlqGew
What we'd basically wana do is use this string to join the host through it's IP address which i hoe we'll figure it out tommorrow. If you want to use the string the user passed it it is stored in the class UserInput. I've added a new method to the Game and some extra Enums that you can probably change becasue they look kinda messy right now.